Rename user to identity
