+++ server/src/article/article.controller.ts
@@ -12,9 +12,9 @@ import {
 import { ArticleService } from './article.service';
 import { CreateArticleDto } from './dto/CreateArticleDto';
 import { Request } from 'express';
-import { JWTGuard } from 'src/user/auth.guard';
+import { JWTGuard } from 'src/identity/auth.guard';
 import { UpdateArticleDto } from './dto/UpdateArticleDto';
-import { LoggedUserGuard } from 'src/user/loggedUser.guard';
+import { LoggedUserGuard } from 'src/identity/loggedUser.guard';
 @Controller('articles')
 export class ArticleController {
   constructor(private readonly articleService: ArticleService) {}
