+++ server/src/follows/entities/Follows.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/user/entities/User';
+import { User } from 'src/identity/entities/User';
 import { Column, Entity, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
 
 @Entity()
