+++ server/src/upload-file/upload-file.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UploadFileService } from './upload-file.service';
 import { UploadFileController } from './upload-file.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/user/entities/User';
+import { User } from 'src/identity/entities/User';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
